requirements.txt
faiss-cpu==1.11.0.post1
sentence-transformers==2.7.0
google-generativeai==0.8.5
beautifulsoup4==4.13.4
playwright==1.54.0
apscheduler==3.11.0
python-dotenv==1.0.1
tqdm==4.66.4
requests==2.31.0
